+++ CHANGES.md
@@ -9,3 +9,6 @@
 
 - [ADD] Windows に DataChannel Observer ブリッジを実装する
   - @zztkm
+
+- [ADD] Windows に EventChannel イベント送出機構を実装する
+  - @zztkm